+++ src/main/kotlin/com/krillsson/sysapi/notifications/NotificationServiceInfo.kt
@@ -4,7 +4,8 @@ data class NotificationServiceInfo(
     val serverId: String,
     val serverName: String,
     val ntfy: NtfyInfo,
-    val webhooks: List<WebhookInfo>
+    val webhooks: List<WebhookInfo>,
+    val mqtt: MqttInfo
 )
 
 data class NtfyInfo(
@@ -16,4 +17,12 @@ data class NtfyInfo(
 data class WebhookInfo(
     val name: String,
     val enabled: Boolean
+)
+
+data class MqttInfo(
+    val enabled: Boolean,
+    val connected: Boolean,
+    val url: String,
+    val stateTopic: String,
+    val eventTopic: String
 )